paypal test : ok. Algorithm of buycart is more efficient, less queries. Need some test but it seems ok.

+++ RailsProject/app/controllers/API/purchases_controller.rb
@@ -21,88 +21,38 @@ module API
     # - +503+ - Error from server
     # 
     def buycart
+      ret = { musics: [], albums: [] }
+
       begin
         if (@security)
-          purchases = { musics: [], albums: [], musicsFromAlbum: [] }
-          cartToDelete = []
+          p = Purchase.new
+          p.user_id = @user_id
+          p.save!
 
-          carts = Cart.where(user_id: @user_id)
-
-          if carts.length > 0
-            p = Purchase.new
-            p.user_id = @user_id
-            p.save!
-            carts.each { |c|
-              cTd = { cart: c, musicToDelete: [], albumToDelete: [] }
-              cartToDelete << cTd
-
-              c.musics.each { |music|
-                pm = PurchasedMusic.new
-                pm.purchase_id = p.id
-                pm.music_id = music.id
-                pm.save!
-                purchases[:musics] << pm
-                cTd[:musicToDelete] << music
-              }
-
-              c.albums.each { |album|
-                pa = PurchasedAlbum.new
-                pa.album_id = album.id
-                pa.save!
-                purchases[:albums] << pa
-                cTd[:albumToDelete] << album
-
-                album.musics.each { |zik|
-                  pm = PurchasedMusic.new
-                  pm.purchase_id = p.id
-                  pm.purchased_album_id = pa.id
-                  pm.music_id = zik.id
-                  pm.save!
-                  purchases[:musicsFromAlbum] << pm
-                }
-              }
-            }
-
-            cartToDelete.each { |c|
-              c[:musicToDelete].each { |music|
-                c[:cart].musics.delete(music)
-              }
-              c[:albumToDelete].each { |album|
-                c[:cart].albums.delete(album)
-              }
-              c[:cart].destroy
-            }
-
-            ret = { musics: [], albums: [] }
-
-            purchases[:musics].each { |p|
-              ret[:musics] << p.as_json(:include => {:music => { only: Music.miniKey } })
-            }
-            purchases[:albums].each { |p|
-              ret[:albums] << p.as_json(:include => {:album => { only: Album.miniKey } })
+          cart = Cart.eager_load(albums: { musics: {} }).where(carts: { user_id: @user_id })
+          list = p.buyCart(cart, false, true)
+          if (list == false)
+            p.destroy
+            @returnValue = { content: nil }
+            codeAnswer 202
+          else
+            list.each { |item|
+              if (item.is_a?(Music))
+                ret[:musics] << p.as_json(:include => {:album => { only: Album.miniKey } })
+              else
+                ret[:albums] << p.as_json(:include => {:musics => { only: Music.miniKey } })
+              end
             }
 
             @returnValue = { content: ret }
             codeAnswer 201
             defineHttp :created
-          else
-            @returnValue = { content: nil }
-            codeAnswer 202
           end
         else
           codeAnswer 500
           defineHttp :forbidden
         end
       rescue
-        purchases[:musics].each { |p|
-          p.destroy
-        }
-        purchases[:albums].each { |p|
-          p.destroy
-        }
-        purchases[:musicsFromAlbum].each { |p|
-          p.destroy
-        }
         codeAnswer 504
         defineHttp :service_unavailable
       end
